optimize:Move TraceDashboardData to mapping package, enhance semantics
